+++ tests/ui/invalid_need_module_arg_position.rs
@@ -2,7 +2,7 @@ use pyo3::prelude::*;
 
 #[pymodule]
 fn module(_py: Python, m: &PyModule) -> PyResult<()> {
-    #[pyfn(m, "with_module", pass_module)]
+    #[pyfn(m, pass_module)]
     fn fail(string: &str, module: &PyModule) -> PyResult<&str> {
         module.name()
     }